Refactored to specify *bulk* download
