+++ src/_scss/pages/bulkDownload/_downloadForm.scss
@@ -25,6 +25,7 @@
             margin-bottom: rem(15);
         }
         .radio {
+            margin-top: rem(20);
             input {
                 margin-right: rem(5);
             }
