+++ src/_scss/pages/bulkDownload/bulkDownloadPage.scss
@@ -1,4 +1,4 @@
-.usa-da-download-page {
+.usa-da-bulk-download-page {
     @import "all";
     @import "layouts/search/search";
     @import "./header/header";
