+++ src/js/redux/reducers/bulkDownload/bulkDownloadReducer.js
